requirements_nlu.txt
wheel
boto3
requests
tqdm
sentence-transformers